Ticket #318 : PD Request - sequence counter not incremented
Because the PD_ELE_T entry for PD Requests is removed from the PD Send Queue as soon as it is sent, hence for maintaining the last used sequence counter, a list of last sequence counter used per each comId for PD Requests is maintained. The pointer to the first element in the list is added in the TRDP_SESSION struct. Every time a new PD request is to be sent, the last sequence counter used for that comId is fetched from this list.

git-svn-id: https://svn.code.sf.net/p/tcnopen/trdp/trunk/trdp@2162 3b5a3598-5f4e-4449-9e63-bd40438bfec0